OpenWeatherMapController Unit Tests - Test correct response is returned and decoded successfully [x] All tests passing
